modified with suspense wrapper
