+++ frontend-ui/src/app/workouts/edit/deleteExercise/page.jsx
@@ -1,29 +1,12 @@
-"use client"
-import { useEffect } from 'react';
-import { useSearchParams, useRouter } from 'next/navigation'; // Import from 'next/navigation'
-import { useAtom } from 'jotai';
-import { editWorkoutAtom, removeExerciseAtom  } from '../../../../../utility/editExerciseAtom'; // Import removeExerciseAtom
+import React, { Suspense } from 'react';
+import DeleteExercisePage from '../../../../components/workouts/DeleteExercisePage';
 
-const DeleteExercisePage = () => {
-  const searchParams = useSearchParams();
-  const id = searchParams.get('id'); // Access the query parameter
-  const workoutId = searchParams.get('wId');
-  const router = useRouter();
-  const [workout, setWorkout] = useAtom(editWorkoutAtom);
-
-  useEffect(() => {
-    if (id) {
-      console.log(id);
-      setWorkout((prevWorkout) => ({
-        ...prevWorkout,
-        exerciseIds: prevWorkout.exerciseIds.filter(exerciseId => exerciseId !== id),
-      }));
-      removeExerciseAtom(id); // Remove the exercise atom
-      router.push(`/workouts/edit/${workoutId}`); // Navigate back to the previous page
-    }
-  }, [id, setWorkout, router]);
-
-  return; //dont return a page
+const DeleteExercisePageWrapper = () => {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <DeleteExercisePage />
+    </Suspense>
+  );
 };
 
-export default DeleteExercisePage;
+export default DeleteExercisePageWrapper;
